% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_simfun.R
\name{check_input_simfun}
\alias{check_input_simfun}
\title{Cointegration test for settings of large N and T}
\usage{
check_input_simfun(N, tau, stat_value, k, r, fin_sample_corr, sim_num)
}
\arguments{
\item{N}{a number representing the number of time series we want to simulate in the system}

\item{tau}{a number representing the length of the time series we want to simulate in the system}

\item{stat_value}{the test statistic value we want to calculate p-value based on}

\item{k}{The number of lags we wish to employ in the VECM form (default: k=1)}

\item{r}{The number of cointegrating relationships we impose on the H1 hypothesis (default: r=1)}

\item{sim_num}{The number of simulation we wish to run.}
}
\description{
This is an internal function that checks the validity of the inputs of the sim_function function.
}
\keyword{internal}
