% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{largevar}
\alias{largevar}
\title{Cointegration test for settings of large N and T}
\usage{
largevar(data, k = 1, r = 1, fin_sample_corr = FALSE, plot_output = TRUE)
}
\arguments{
\item{data}{a numeric matrix where columns contain the individual time series that will be examined for presence of cointegrating relationships}

\item{k}{The number of lags we wish to employ in the VECM form (default: k=1)}

\item{r}{The number of cointegrating relationships we impose on the H1 hypothesis (default: r=1)}

\item{fin_sample_corr}{A boolean variable indicating whether we wish to employ finite sample correction on our test statistic. Default is false}

\item{plot_output}{A boolean variable indicating whether we wish to generate the distribution of the eigenvalues (default: TRUE)}
}
\value{
A list that contains the test statistic, the upper and lower bounds of the support of the measure of the Wachter distribution, the eigenvalues of our sample matrix, and a table of significance for the test statistic.
}
\description{
Runs the Bykhovskaya-Gorin test for cointegration. Paper can be found at: https://doi.org/10.48550/arXiv.2202.07150
}
\examples{
result <- largevar(data=data,k=1,r=1,plot_output=FALSE);
}
