% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{largevar_scel}
\alias{largevar_scel}
\title{Internal skeleton function of cointegration test for simfun function}
\usage{
largevar_scel(data, k, r, fin_sample_corr)
}
\arguments{
\item{data}{a numeric matrix where columns contain the individual time series that will be examined for presence of cointegrating relationships}

\item{k}{The number of lags we wish to employ in the VECM form (default: k=1)}

\item{r}{The number of cointegrating relationships we impose on the H1 hypothesis (default: r=1)}

\item{fin_sample_corr}{A boolean variable indicating whether we wish to employ finite sample correction on our test statistic. Default is false.}
}
\value{
The test statistic.
}
\description{
This is the "skeleton" version of the largevar function in the package. It is called within the sim_function function to make runtime faster. For the actual cointegration test, use the largevar function.
}
\keyword{internal}
