% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentiles.R
\docType{data}
\name{percentiles}
\alias{percentiles}
\title{Quantiles for the limiting distribution of the test}
\format{
A data frame with 99 rows and 11 variables:
}
\source{
Calculated through own simulations (see details in Section 4 of vignette).
}
\description{
A data frame containing the simulated quantiles for the test statistic used in
 the largevar function. More details about how these simulations were conducted
 can be found in Section 4 of the vignette.
}
