% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_tables.R
\name{quantile_tables}
\alias{quantile_tables}
\title{Creates the quantile table output for largevar function}
\usage{
quantile_tables(r = 1)
}
\arguments{
\item{r}{Which partial sum the quantile table should be returned for. (Only r<=10 is available.) Default is r=1.}
}
\value{
A numeric matrix.
}
\description{
Outputs the quantile tables from the package's corresponding vignette.
}
\examples{
quantile_tables(r=3)
}
