% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_p100_price_adj.R
\docType{data}
\name{s_p100_price_adj}
\alias{s_p100_price_adj}
\title{Stock price data for example in vignette}
\format{
A data frame with 522 rows and 93 variables:
}
\source{
Refer to the data source used in: A. Bykhovskaya and V. Gorin. Cointegration in large vars. Annals of Statistics, 2022.
}
\description{
A data frame containing weekly S&P100 prices over ten years: 01.01.2010 - 01.01.2020,
 The S&P100 includes 101 leading U.S. stocks of which 92 were collected here.
}
