% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{class}
\name{stat_test-class}
\alias{stat_test-class}
\title{Define a class for the largervar() output list
# Define the class}
\description{
Define a class for the largervar() output list
# Define the class
}
