% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_simfun.R
\name{check_input_simfun}
\alias{check_input_simfun}
\title{Cointegration test for settings of large N and T}
\usage{
check_input_simfun(data, k, r, sim_num)
}
\arguments{
\item{data}{a numeric matrix where columns contain the individual time series that will be examined for presence of cointegrating relationships}

\item{k}{The number of lags we wish to employ in the VECM form (default: k=1)}

\item{r}{The number of cointegrating relationships we impose on the H1 hypothesis (default: r=1)}

\item{sim_num}{The number of simulation we wish to run.}
}
\description{
This is an internal function that checks the validity of the inputs of the sim_function function.
}
\keyword{internal}
