% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{largevar_scel}
\alias{largevar_scel}
\title{Cointegration test for settings of large N and T}
\usage{
largevar_scel(data, k = 1, r = 1)
}
\arguments{
\item{data}{a numeric matrix where columns contain the individual time series that will be examined for presence of cointegrating relationships}

\item{k}{The number of lags we wish to employ in the VECM form (default: k=1)}

\item{r}{The number of cointegrating relationships we impose on the H1 hypothesis (default: r=1)}
}
\value{
The test statistic.
}
\description{
This is the "skeleton" version of the largevar function in the package. It is called within the sim_function function to make runtime faster. For the actual cointegration test, use the largevar function.
}
