% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mains.R
\name{sim_function}
\alias{sim_function}
\title{Cointegration test for settings of large N and T}
\usage{
sim_function(data, k = 1, r = 1, sim_num = 500)
}
\arguments{
\item{data}{a numeric matrix where columns contain the individual time series that will be examined for presence of cointegrating relationships}

\item{k}{The number of lags we wish to employ in the VECM form (default: k=1)}

\item{r}{The number of cointegrating relationships we impose on the H1 hypothesis (default: r=1)}

\item{sim_num}{The number of simulations that the function conducts for H_0 (the function is slow!). Default is 500.}
}
\value{
A list that contains the simulation values, the empirical percentage (realizations larger than the test statistic for the original data) and a histogram.
}
\description{
Runs a simulation on the H0 for the Bykhovskaya-Gorin test for cointegration and returns an empirical p-value. Paper can be found at: https://doi.org/10.48550/arXiv.2202.07150
}
\examples{
result <- sim_function(data,k=1,r=1,sim_num=20)
}
